pyproject.toml
# Poetry pyproject.toml: https://python-poetry.org/docs/pyproject/

[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "doh"
version = "0.3.0"
description = "Make docker interactive usage less painful"
readme = "README.md"
authors = [
  "Vsevolod Poletaev <poletaev.va _at_ gmail.com>"
]
license = "GNU GPL v3.0"
repository = "https://github.com/hexfaker/doh"
homepage = "https://github.com/hexfaker/doh"

# Keywords description https://python-poetry.org/docs/pyproject/#keywords
keywords = []  # Update me

include = [
  {path="doh/bin/ssh-server", format="wheel"}
]

# Pypi classifiers: https://pypi.org/classifiers/
classifiers = [  # Update me
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "Operating System :: OS Independent",
  "Topic :: Software Development :: Libraries :: Python Modules",
]

[tool.poetry.scripts]
# Entry points for the package https://python-poetry.org/docs/pyproject/#scripts
"doh" = "doh.__main__:app"


[tool.poetry.dependencies]
python = "^3.8"
typer = "^0.6.1"
pydantic = "^1.7.3"
toml = "^0.10.2"
rich = "^12.3.0"
platformdirs = "^2.5.2"
tomlkit = "^0.11.0"
python-dotenv = "^1.0.0"

[tool.poetry.dev-dependencies]
darglint = "^1.5.8"
isort = "^5.7.0"
pyupgrade = "^3.1.0"
black = "^22.3.0"
mypy = "^1.2"
safety = "^2.0.0"
pytest = "^7.1.2"
pylint = "^2.13.7"
pydocstyle = "^6.1.1"
pre-commit = "^2.9.3"
types-toml = "^0.10.6"

[tool.black]
# https://github.com/psf/black
line-length = 80
target-version = ["py38"]

[tool.isort]
# https://github.com/timothycrosley/isort/
known_typing = "typing,types,typing_extensions,mypy,mypy_extensions"
sections = "FUTURE,TYPING,STDLIB,THIRDPARTY,FIRSTPARTY,LOCALFOLDER"
include_trailing_comma = true
default_section = "FIRSTPARTY"
multi_line_output = 3
indent = 4
force_grid_wrap = 0
use_parentheses = true
line_length = 80
